cleaning up movement and adding matrix Options for better performance

+++ src/settings.rs
@@ -5,7 +5,7 @@ SETTINGS
 pub static VSYNC: bool = false;
 
 use std::time::Duration;
-pub static TICKS_PER_SECOND: u64 = 25;
-pub static MCS_PER_UPDATE: Duration = Duration::from_micros(1000000 / TICKS_PER_SECOND);
+pub static TICKS_PER_SECOND: f32 = 25.0;
+pub static MCS_PER_UPDATE: Duration = Duration::from_micros(1000000 / TICKS_PER_SECOND as u64);
 
 pub static WINDOW_LABEL: &str = "punt the weak";
